Move prepare_all_load_points out of object module into preprocessing
