+++ flexible_load_analysis/init/preprocessing.py
@@ -1,5 +1,14 @@
+import datetime as dt
+
 import numpy as np
 
+from ..modelling import modelling
+from .. import utilities
+
+"""Module for performing various timeseries preprocessing steps.
+The module functionality is designed to be pipeline-able, in that
+a choice of preprocessing methods may be run back to back. 
+"""
 
 def remove_nan_and_none_datapoints(ts_data):
     """Removes datapoints containing NaN or None.
@@ -231,3 +240,60 @@ def preprocess_data(dict_preprocessing_config, dict_data_ts):
 
     print("Successfully completed all preprocessing steps")
     return dict_data_ts
+
+
+# TODO: Messy, not expandable, uneccesary.
+# Weird to have both preprocessing and prepare_all_loads, which does preprocessing with prework
+def prepare_all_loads(dict_config, dict_data):
+    """Prepares nodes based on input data and config.
+    Parameters
+    ----------
+    dict_config : dict
+        Configuration-file.
+    dict_data : dictionary of measured loads and temperature.
+    """
+    print("Preparing common data...")
+    date_start = dt.date.fromisoformat(
+        dict_config["data"]["load_measurements"]["first_date_iso"])
+    date_end = dt.date.fromisoformat(
+        dict_config["data"]["load_measurements"]["last_date_iso"])
+
+    if dict_config["preprocessing"]["correct_for_temperature"]:
+        ts_temperature_historical = utilities.get_first_value_of_dictionary(
+            dict_data["temperature_measurements"])
+        ts_temperature_historical = remove_nan_and_none_datapoints(
+            ts_temperature_historical)
+        dict_daily_normal_temperature = compute_daily_historical_normal(
+            ts_temperature_historical)
+        dict_temperature_n_day_average = create_n_day_average_dict(
+            ts_temperature_historical,
+            date_start, date_end,  n=3)
+
+    print("Preparing all loads in network...")
+    # Preprocessing and potential modelling of every load-point
+    dict_loads_ts = {}
+    for str_node_ID in dict_data["load_measurements"]:
+        print("--------------------")
+        print("Preparing load-point", str_node_ID + "...")
+
+        dict_node_ts = {}
+        dict_node_ts["load_measurements"] = dict_data["load_measurements"][str_node_ID]
+        if dict_config["preprocessing"]["correct_for_temperature"]:
+            dict_node_ts["normal_temperature"] = dict_daily_normal_temperature
+            dict_node_ts["n-day_average_temperature"] = dict_temperature_n_day_average
+
+        print("Preprocessing", str_node_ID + "...")
+        dict_node_ts = preprocess_data(
+            dict_config["preprocessing"], dict_node_ts)
+
+        if dict_config["modelling"]["perform_modelling"]:
+            print("Modelling based on dataset", str_node_ID + "...")
+            dict_model = modelling.model_load(
+                dict_config["modelling"], dict_node_ts)
+            dict_loads_ts[str_node_ID] = dict_model["load"]
+        
+        dict_loads_ts[str_node_ID] = dict_node_ts["load"]
+
+    print("--------------------")
+    print("Successfully prepared all load-points")
+    return dict_loads_ts
